game-core/implement a spell script framework

+++ service-world/src/main/java/com/github/azeroth/game/domain/map/PositionFullTerrainStatus.java
@@ -7,7 +7,7 @@ import lombok.Data;
 
 @Data
 public class PositionFullTerrainStatus {
-    private short areaId;
+    private int areaId;
     private float floorZ;
     private boolean outdoors;
     private EnumFlag<ZLiquidStatus> liquidStatus;
